GCP-TF/modules/batch_job/functions/requirements.txt
boto3
httpx
requests
google-cloud-batch
google-cloud-pubsub
strands-agents
strands-agents-tools